+++ src/components/PrivateFeatures/ChatBot/components/layouts/main/index.js
@@ -1,0 +1,1 @@
+export { default } from "./Main";
