+++ src/components/PrivateFeatures/ChatBot/components/modules/conversations/Conversation/index.js
@@ -1,0 +1,1 @@
+export { default } from "./Conversation";
