+++ src/components/PrivateFeatures/ChatBot/components/modules/conversations/Conversations/Conversations.jsx
@@ -1,13 +1,13 @@
 import React, { useRef, useState } from "react";
 
-import styled from "styled-components";
+import Conversation from "../Conversation/Conversation";
+import CreateConversation from "../CreateConversation";
 
-import Conversation from "./Conversation";
-import CreateConversation from "./CreateConversation";
+import Wrapper from "./style";
 
-import { EditIcon, LeftIcon, RightIcon } from "../../icons";
+import { EditIcon, LeftIcon, RightIcon } from "../../../icons";
 
-import { useConversation } from "../../../hooks";
+import { useConversation, useStorage } from "../../../../hooks";
 
 const Conversations = () => {
   const {
@@ -17,7 +17,10 @@ const Conversations = () => {
     deleteConversation,
     setCurrentConversation,
   } = useConversation();
-  const [isCollapsed, setIsCollapsed] = useState(false);
+  const [isCollapsed, setIsCollapsed] = useStorage(
+    "CHATBOT-SIDEBAR-STATE",
+    false
+  );
 
   const toggleCollapse = () => {
     setIsCollapsed((prev) => !prev);
@@ -87,67 +90,5 @@ const Conversations = () => {
     </>
   );
 };
-
-const Wrapper = styled.div`
-  height: 100%;
-  .conversations {
-    height: 100%;
-    width: 30vw;
-    max-width: 300px;
-    display: flex;
-    flex-direction: column;
-    gap: 0.5em;
-    transition: transform 0.2s ease-in;
-    &.slide-left {
-      z-index: 1;
-      position: absolute;
-      transform: translateX(calc(-100% - 2em));
-    }
-  }
-  .empty-conversation {
-    text-align: center;
-    color: var(--chatbot-grey);
-  }
-  .action-button {
-    position: relative;
-    background-color: white;
-    padding: 0.75em;
-    color: var(--chatbot-grey);
-    border-radius: 0.25em;
-    display: flex;
-    justify-content: center;
-    border: 1px solid var(--chatbot-grey);
-    align-items: center;
-  }
-  display: flex;
-  flex-direction: column;
-  gap: 0.5em;
-  .create-new-chat {
-    background-color: white;
-    display: flex;
-    font-size: 1.1em;
-    padding: 0.5em 1.25em;
-    font-weight: bold;
-    justify-content: space-between;
-    align-items: center;
-    border-radius: 1em;
-    border: 1px solid var(--chatbot-grey);
-    .edit-button {
-      color: var(--chatbot-grey);
-    }
-  }
-  .conversation-list {
-    border-radius: 1em;
-    background-color: white;
-    padding: 0.5em;
-    overflow: hidden auto;
-    flex: 1;
-    border: 1px solid var(--chatbot-grey);
-    .conversation.active {
-      color: white;
-      background-color: var(--chatbot-primary);
-    }
-  }
-`;
 
 export default Conversations;
