+++ src/components/PrivateFeatures/ChatBot/components/modules/conversations/Conversations/index.js
@@ -1,0 +1,1 @@
+export { default } from "./Conversations";
