+++ src/components/PrivateFeatures/ChatBot/components/modules/conversations/index.js
@@ -1,1 +1,1 @@
-export { default as SideBar } from "./Conversations";
+export { default as SideBar } from "./Conversations/Conversations";
