+++ src/components/PrivateFeatures/ChatBot/hooks/index.js
@@ -1,1 +1,2 @@
 export { default as useConversation } from "./useConversation";
+export { default as useStorage } from "./useStorage";
